algunos cambiso pero en proceso de cambiar consultorios y calendariso

+++ src/components/tratamientos/tratamientosController.js
@@ -20,7 +20,8 @@ export const getTratamientoById = async (req, res) => {
 };
 
 export const createTratamiento = async (req, res) => {
-    const { nombre, descripcion, duracion, costo } = req.body;
+    const { nombre, descripcion, duracion, costo, color } = req.body;
+
 
     // Validar que todos los campos estén presentes
     if (!nombre || !descripcion || !duracion || !costo) {
@@ -29,8 +30,8 @@ export const createTratamiento = async (req, res) => {
 
     try {
         const result = await pool.query(
-            'INSERT INTO tratamientos (nombre, descripcion, duracion, costo) VALUES ($1, $2, $3, $4) RETURNING *',
-            [nombre, descripcion, duracion, costo]
+            'INSERT INTO tratamientos (nombre, descripcion, duracion, costo, color) VALUES ($1, $2, $3, $4, $5) RETURNING *',
+            [nombre, descripcion, duracion, costo, color]
         );
         res.status(201).json(result.rows[0]);
     } catch (error) {
@@ -41,9 +42,12 @@ export const createTratamiento = async (req, res) => {
 
 export const updateTratamiento = async (req, res) => {
     const id = req.params.id_tratamiento;
-    const { nombre, descripcion } = req.body;
+    const { nombre, descripcion, duracion, costo, color } = req.body;
+
+    console.log("Datos recibidos:", id, nombre, descripcion, duracion, costo);
+
     try {
-        const result = await pool.query('UPDATE tratamientos SET nombre = $1, descripcion = $2 WHERE id_tratamiento = $3 RETURNING *', [nombre, descripcion, id]);
+        const result = await pool.query('UPDATE tratamientos SET nombre = $1, descripcion = $2 , duracion = $3, costo = $4, color = $5 WHERE id_tratamiento = $6 RETURNING *', [nombre, descripcion, duracion, costo, color, id]);
         res.status(200).json(result.rows[0]);
     } catch (error) {
         res.status(500).json({ message: 'Error al actualizar tratamiento', error: error.message });
@@ -52,7 +56,17 @@ export const updateTratamiento = async (req, res) => {
 
 export const deleteTratamiento = async (req, res) => {
     const id = req.params.id_tratamiento;
+
     try {
+
+        const asociaciones = await pool.query('SELECT *   FROM consultorio_tratamiento WHERE id_tratamiento = $1', [id]);
+
+        if (asociaciones.rows.length > 0) {
+            return res.status(400).json({
+                message: 'No se puede eliminar un tratamiento asociado a un consultorio',
+                details: asociaciones.rows, // Opcional: proporciona detalles de las asociaciones
+            });
+        }
         const result = await pool.query('DELETE FROM tratamientos WHERE id_tratamiento = $1 RETURNING *', [id]);
         res.status(200).json(result.rows[0]);
     } catch (error) {
